requirements.txt
openenv[core]>=0.2.0
fastapi>=0.115.0
uvicorn>=0.24.0
gymnasium>=0.28
httpx>=0.27
openai>=1.0
pydantic>=2.8
pyyaml>=6.0
